% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_by_git}
\alias{create_by_git}
\title{Create in bulk by searching for git repos}
\usage{
create_by_git(proj_dir, workbch_dir = getOption("workbch.home"))
}
\arguments{
\item{proj_dir}{where to look for git repositories}

\item{workbch_dir}{where to store workbench files}
}
\description{
Create in bulk by searching for git repos
}
