% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_jobs_by_git}
\alias{create_jobs_by_git}
\title{Create in bulk by searching for git repos}
\usage{
create_jobs_by_git(dir, owner, status = "active", priority = 1,
  deadline = NA, hidden = FALSE)
}
\arguments{
\item{dir}{where to look for git repositories}

\item{owner}{the owner for created jobs (a nick name)}

\item{status}{the status for created jobs}

\item{priority}{the priority of created jobs}

\item{deadline}{the deadline of created jobs}

\item{hidden}{the visibility of created jobs}
}
\description{
Create in bulk by searching for git repos
}
\examples{
\dontrun{

create_jobs_by_git(dir = "~/my_projects", owner = "me")
}

}
