% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_job}
\alias{delete_job}
\title{Delete a job}
\usage{
delete_job(jobname)
}
\arguments{
\item{jobname}{name of the job to delete}
}
\description{
Delete a job
}
\details{
Deletes a job from the JSON file.
}
\examples{
\dontrun{

delete_job("myjob")
}
}
