% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_url}
\alias{delete_url}
\title{Delete a URL from a job}
\usage{
delete_url(name, site)
}
\arguments{
\item{name}{the job from which the URL should be deleted}

\item{site}{the name of the site to which the URL should be deleted}
}
\description{
Delete a URL from a job
}
\examples{
\dontrun{

delete_url("myjob", "github")
}
}
