% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goto.R
\name{goto_job}
\alias{goto_job}
\title{Navigate to a job}
\usage{
goto_job(jobname)
}
\arguments{
\item{jobname}{name of job to open}
}
\description{
Navigate to a job
}
\examples{
\dontrun{

goto_job("myjob")
}
}
