% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goto.R
\name{goto_project}
\alias{goto_project}
\title{Navigate to the RStudio project linked with a job}
\usage{
goto_project(name)
}
\arguments{
\item{name}{name of job to open}
}
\description{
Navigate to the RStudio project linked with a job
}
\examples{
\dontrun{

goto_project("myjob")
}
}
