% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{initialise_by_git}
\alias{initialise_by_git}
\title{Initialise by searching for git repos}
\usage{
initialise_by_git(proj_dir, workbch_dir = getOption("workbch.home"))
}
\arguments{
\item{proj_dir}{where to look for git repositories}

\item{workbch_dir}{where to store workbench files}
}
\description{
Initialise by searching for git repos
}
