% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{job_add_note}
\alias{job_add_note}
\title{Append a note to a job}
\usage{
job_add_note(name, note)
}
\arguments{
\item{name}{the job to which the note should be added}

\item{note}{the text of the note}
}
\description{
Append a note to a job
}
\examples{
\dontrun{

job_add_note("myjob", "susan wanted me to notify her when done")
}
}
