% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_task.R
\name{job_addtask}
\alias{job_addtask}
\title{Attach a new task to job}
\usage{
job_addtask(jobname, description, owner = NULL, status = "active",
  priority = NULL, deadline = NULL, hidden = NULL)
}
\arguments{
\item{jobname}{name of the job the task attaches to}

\item{description}{brief description of the task}

\item{owner}{should be a name or a nickname (defaults to job owner)}

\item{status}{should be "active" (default), "inactive", "complete", "abandoned"}

\item{priority}{numeric (default is to match the job)}

\item{deadline}{a date (default is to match the job)}

\item{hidden}{hide the task (default is to match the job)}
}
\description{
Attach a new task to job
}
