% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_browse.R
\name{job_browse}
\alias{job_browse}
\title{Navigate to a URL linked to a job}
\usage{
job_browse(site, jobname = NULL)
}
\arguments{
\item{site}{label denoting the site (e.g., "github")}

\item{jobname}{name of the job}
}
\description{
Navigate to a URL linked to a job
}
