% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{job_browse_url}
\alias{job_browse_url}
\title{Open a url linked to a job}
\usage{
job_browse_url(name, site)
}
\arguments{
\item{name}{name of the project}

\item{site}{label denoting the site (e.g., "github")}
}
\description{
Open a url linked to a job
}
