% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{job_create}
\alias{job_create}
\title{Create a new job}
\usage{
job_create(jobname = NULL, description = NULL, owner = NULL,
  status = NULL, priority = NULL, tags = NULL, path = NULL)
}
\arguments{
\item{jobname}{name of the job to create}

\item{description}{brief description of the job}

\item{owner}{should be a name or a nickname}

\item{status}{should be "active", "inactive", "complete", "abandoned", "masked"}

\item{priority}{numeric}

\item{tags}{character vector of tags}

\item{path}{path to the job home directory}
}
\description{
Create a new job
}
