% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_project.R
\name{job_delete}
\alias{job_delete}
\title{Delete a job}
\usage{
job_delete(name)
}
\arguments{
\item{name}{name of the job to delete}
}
\description{
Delete a job
}
\details{
Deletes a job from the JSON file. At the moment, it does so without
asking for the user to confirm, so be careful.
}
\examples{
\dontrun{

job_delete("myjob")
}
}
