% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{job_delete_note}
\alias{job_delete_note}
\title{Delete a note from a job}
\usage{
job_delete_note(name, id)
}
\arguments{
\item{name}{the job from which the note should be deleted}

\item{id}{the id number assigned to the note}
}
\description{
Delete a note from a job
}
\examples{
\dontrun{

job_delete_note("myjob", 2)
}
}
