% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{job_edit}
\alias{job_edit}
\title{Edit a job}
\usage{
job_edit(name, ...)
}
\arguments{
\item{name}{name of the project to edit}

\item{...}{expressions to be evaluated in the job}
}
\description{
Edit a job
}
