% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_project.R
\name{job_edit_url}
\alias{job_edit_url}
\title{Add or edit urls associated with a job}
\usage{
job_edit_url(name, site, link)
}
\arguments{
\item{name}{name of the project to edit}

\item{site}{string with the site nickname (e.g., "github")}

\item{link}{string with the link to the site}
}
\description{
Add or edit urls associated with a job
}
\details{
The role of \code{job_edit_url()} is to make it a easier to
change a URL associated with a job.
}
\examples{
\dontrun{

job_edit_url("myjob", "github", "https://github.com/myusername/myrepo")

}
}
