% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_gitrep.R
\name{job_gitrep}
\alias{job_gitrep}
\title{Report the git status of all jobs}
\usage{
job_gitrep(show_clean = FALSE)
}
\arguments{
\item{show_clean}{should clean repos be included?}
}
\value{
A tibble
}
\description{
Report the git status of all jobs
}
