% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_gitreport.R
\name{job_gitreport}
\alias{job_gitreport}
\title{Report the git status of all jobs}
\usage{
job_gitreport(show_clean = FALSE)
}
\arguments{
\item{show_clean}{should clean repos be included?}
}
\value{
A tibble
}
\description{
Report the git status of all jobs
}
