% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_modify.R
\name{job_modify}
\alias{job_modify}
\title{Modify the properties of a job}
\usage{
job_modify(jobname = NULL, newname = NULL, description = NULL,
  owner = NULL, status = NULL, priority = NULL, path = NULL,
  tags = NULL, url = NULL, delete = FALSE)
}
\arguments{
\item{jobname}{the current name for the job}

\item{newname}{the new name}

\item{description}{the new description}

\item{owner}{nick name for the new owner}

\item{status}{the new status}

\item{priority}{the new priority}

\item{path}{the new path to the job folder}

\item{tags}{string with tags (uses | as separator)}

\item{url}{string specifying url}

\item{delete}{should this job be deleted (default = FALSE)}
}
\description{
Modify the properties of a job
}
