% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_task}
\alias{make_task}
\title{Make a task and attach to a job}
\usage{
make_task(description, jobname = NULL, owner = NULL,
  status = "active", priority = NULL, deadline = NULL,
  hidden = NULL)
}
\arguments{
\item{description}{brief description of the task}

\item{jobname}{name of the job the task attaches to}

\item{owner}{should be a name or a nickname (defaults to job owner)}

\item{status}{should be "active" (default), "inactive", "complete", "abandoned"}

\item{priority}{numeric (default is to match the job)}

\item{deadline}{a date (default is to match the job)}

\item{hidden}{hide the task (default is to match the job)}
}
\description{
Make a task and attach to a job
}
