% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{people_add}
\alias{people_add}
\title{Add a new person}
\usage{
people_add(name, nickname)
}
\arguments{
\item{name}{the name of the person}

\item{nickname}{a nickname for the person}
}
\description{
Add a new person
}
