% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R, R/goto.R
\name{projectr_home}
\alias{projectr_home}
\title{Get or set the location of projectr files}
\usage{
projectr_home(path = NULL)

projectr_home(path = NULL)
}
\arguments{
\item{path}{Path to the folder}

\item{path}{Path to the folder}
}
\value{
Path to the folder

Path to the folder
}
\description{
Get or set the location of projectr files

Get or set the location of projectr files
}
