% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_deadline}
\alias{set_deadline}
\title{Set the deadline for a job}
\usage{
set_deadline(date, jobname = NULL)
}
\arguments{
\item{date}{character string to be parsed by lubridate::dmy}

\item{jobname}{name of job(s) to be edited}
}
\description{
Set the deadline for a job
}
