% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_description}
\alias{set_description}
\title{Sets the description of a job}
\usage{
set_description(jobname, description)
}
\arguments{
\item{jobname}{Name of the job}

\item{description}{String specifying the description}
}
\description{
Sets the description of a job
}
