% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_hidden}
\alias{set_hidden}
\title{Sets the visibility of a job}
\usage{
set_hidden(jobname, hidden)
}
\arguments{
\item{jobname}{Name of the job}

\item{hidden}{Logical}
}
\description{
Sets the visibility of a job
}
