% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_job.R
\name{set_job}
\alias{set_job}
\alias{set_job_name}
\alias{set_job_description}
\alias{set_job_status}
\alias{set_job_owner}
\alias{set_job_priority}
\alias{set_job_path}
\alias{set_job_team}
\alias{set_job_url}
\alias{set_job_tag}
\title{Set the properties of an existing job}
\usage{
set_job_name(jobname, newname)

set_job_description(jobname, description)

set_job_status(jobname, status)

set_job_owner(jobname, owner)

set_job_priority(jobname, priority)

set_job_path(jobname, path)

set_job_team(jobname, add_team = NULL, remove_team = NULL)

set_job_url(jobname, site, link)

set_job_tag(jobname, add_tag = NULL, remove_tag = NULL)
}
\arguments{
\item{jobname}{the current name for the job}

\item{newname}{the new name}

\item{description}{the new description}

\item{status}{the new status}

\item{owner}{nick name for the new owner}

\item{priority}{the new priority}

\item{path}{the new path to the job folder}

\item{add_team}{character vector of nick names to add to the team}

\item{remove_team}{character vector of nick names to remove from the team}

\item{site}{string with the site nickname (e.g., "github")}

\item{link}{string with the link to the site}

\item{add_tag}{character vector of tags to add to jobs}

\item{remove_tag}{character vector of tags to remove from job}
}
\description{
Set the properties of an existing job
}
