% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_job_description}
\alias{set_job_description}
\title{Sets the description of a job}
\usage{
set_job_description(description, jobname = NULL)
}
\arguments{
\item{description}{String specifying the description}

\item{jobname}{Name of the job}
}
\description{
Sets the description of a job
}
