% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_job_hidden}
\alias{set_job_hidden}
\title{Sets the visibility of a job}
\usage{
set_job_hidden(hidden, jobname = NULL)
}
\arguments{
\item{hidden}{Logical}

\item{jobname}{Name of the job}
}
\description{
Sets the visibility of a job
}
