% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_job_path}
\alias{set_job_path}
\title{Sets the path of a job}
\usage{
set_job_path(path, jobname = NULL)
}
\arguments{
\item{path}{The path to the job folder}

\item{jobname}{Name of the job}
}
\description{
Sets the path of a job
}
