% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_jobname}
\alias{set_jobname}
\title{Sets the name of a job}
\usage{
set_jobname(from, to)
}
\arguments{
\item{from}{the current name for the job}

\item{to}{the new name for the job}
}
\description{
Sets the name of a job
}
