% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_note}
\alias{set_note}
\title{Set a note linked to a job}
\usage{
set_note(note, jobname = NULL)
}
\arguments{
\item{note}{the text of the note}

\item{jobname}{the job to which the note should be added}
}
\description{
Set a note linked to a job
}
\examples{
\dontrun{

set_note("myjob", "susan wanted me to notify her when done")
}
}
