% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_owner}
\alias{set_owner}
\title{Sets the owner of a job}
\usage{
set_owner(owner, jobname = NULL)
}
\arguments{
\item{owner}{Nick name for the new owner}

\item{jobname}{Name of the job}
}
\description{
Sets the owner of a job
}
