% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_path}
\alias{set_path}
\title{Sets the path of a job}
\usage{
set_path(jobname, path)
}
\arguments{
\item{jobname}{Name of the job}

\item{path}{The path to the job folder}
}
\description{
Sets the path of a job
}
