% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_priority}
\alias{set_priority}
\title{Sets the priority of a job}
\usage{
set_priority(priority, jobname = NULL)
}
\arguments{
\item{priority}{The new priority}

\item{jobname}{Name of the job}
}
\description{
Sets the priority of a job
}
