% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_task.R
\name{set_task}
\alias{set_task}
\alias{set_task_description}
\alias{set_task_status}
\alias{set_task_owner}
\alias{set_task_priority}
\alias{set_task_deadline}
\title{Set the properties of an existing task}
\usage{
set_task_description(ref, description)

set_task_status(ref, status)

set_task_owner(ref, owner)

set_task_priority(ref, priority)

set_task_deadline(ref, deadline)
}
\arguments{
\item{ref}{either a number (id) or a string to be matched against}

\item{description}{brief description of the task}

\item{status}{should be "active" (default), "inactive", "complete", "abandoned"}

\item{owner}{should be a name or a nickname (defaults to job owner)}

\item{priority}{numeric (default is to match the job)}

\item{deadline}{a date (default is to match the job)}
}
\description{
Set the properties of an existing task
}
