% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_team}
\alias{set_team}
\title{Set the members of a team}
\usage{
set_team(name, add = NULL, remove = NULL)
}
\arguments{
\item{name}{name of job to be edited}

\item{add}{character vector of names to add to the team}

\item{remove}{character vector of names to remove from the team}
}
\description{
Set the members of a team
}
\details{
The role of \code{set_team()} is to make it a little easier to
alter the set of names listed in the "team" field of a job. The same task
could be done with \code{set_job()} but it is cumbersome to do so. The
function first appends any names in \code{add} to the team (ignoring
duplicates), and then removes any names listed in \code{remove}. The function
checks all elements of \code{add} and \code{remove} against the table of
known nicknames, and substitutes the full name in place of a nickname.
Note that it is not possible to remove the "owner" of a job using this
method, as this would leave the job without an owner. To remove the owner,
first use \code{set_job()} to change the owner, and then use
\code{set_team()} to remove the former owner from the team.
}
\examples{
\dontrun{

set_team("myjob", add = "hayley")
set_team("myjob", add = c("hayley", "sarah"))
set_team("myjob", remove = c("hayley", "sarah"))

# if "hayley" is the current owner"
set_job("myjob", owner = "sarah") # transfers the ownership to sarah
set_team("myjob", remove = "hayley") # removes hayley entirely
}
}
