% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_workbch_home}
\alias{set_workbch_home}
\title{Set the location of workbch files}
\usage{
set_workbch_home(path = NULL)
}
\arguments{
\item{path}{Path to the folder}
}
\value{
Path to the folder
}
\description{
Set the location of workbch files
}
