% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbch.R
\name{there}
\alias{there}
\title{File paths specified from job home}
\usage{
there(jobname, ...)
}
\arguments{
\item{jobname}{Name of job}

\item{...}{Arguments to be passed to file.path}
}
\description{
File paths specified from job home
}
