% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_git_status}
\alias{view_git_status}
\title{View the git status of jobs}
\usage{
view_git_status(show_hidden = TRUE, show_clean = FALSE)
}
\arguments{
\item{show_hidden}{should hidden jobs be included?}

\item{show_clean}{should clean repos be included?}
}
\value{
A tibble
}
\description{
View the git status of jobs
}
