% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewing.R
\name{view_gitstatus}
\alias{view_gitstatus}
\title{Check the git status of projects}
\usage{
view_gitstatus(show_hidden = FALSE, show_clean = FALSE)
}
\arguments{
\item{show_hidden}{should hidden jobs be included?}

\item{show_clean}{should clean repos be included?}
}
\value{
A tibble
}
\description{
Check the git status of projects
}
