% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewing.R
\name{view_joblist}
\alias{view_joblist}
\title{List jobs}
\usage{
view_joblist(..., show_hidden = FALSE)
}
\arguments{
\item{...}{expression to be passed to dplyr::filter}

\item{show_hidden}{should hidden jobs be included}
}
\description{
List jobs
}
