% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_jobs}
\alias{view_jobs}
\title{View a list of jobs}
\usage{
view_jobs(priority = 1:2, status = c("active", "inactive"),
  owner = NULL, tags = NULL, cols = c("jobname", "owner", "priority",
  "status", "description"))
}
\arguments{
\item{priority}{what priority levels to show in the output}

\item{status}{what status values to show in the output}

\item{owner}{what owner values to show in the output}

\item{tags}{what tags to show in the output}

\item{cols}{what columns to show in the output}
}
\description{
View a list of jobs
}
