% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_notes}
\alias{view_notes}
\title{View notes associated with a job}
\usage{
view_notes(jobname)
}
\arguments{
\item{jobname}{the job}
}
\value{
Invisibly returns a tibble containing columns for the note,
the id number, the creation date, and the job name
}
\description{
View notes associated with a job
}
\details{
Displays all notes associated with a job in order of recency. The
display format is minimal, showing only an id number (to make it easy to
delete notes later) and the text of each note. Notes are shown in
chronological order, with most recent notes at the top of the output
}
\examples{
\dontrun{

view_notes("myjob")
}
}
