% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_paths}
\alias{view_paths}
\title{View the job folder locations known to workbch}
\usage{
view_paths(show_hidden = TRUE)
}
\arguments{
\item{show_hidden}{should hidden jobs be included}
}
\value{
A tibble
}
\description{
View the job folder locations known to workbch
}
