% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_tag}
\alias{view_tag}
\title{View a list of jobs that possess a tag}
\usage{
view_tag(tag, ..., show_hidden = TRUE, invert = FALSE)
}
\arguments{
\item{tag}{what tag to display}

\item{...}{filtering expression to pass to dplyr::filter}

\item{show_hidden}{should hidden jobs be included}

\item{invert}{if TRUE, return jobs without the tag}
}
\description{
View a list of jobs that possess a tag
}
