% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_tasks}
\alias{view_tasks}
\title{View tasks}
\usage{
view_tasks(..., show_hidden = FALSE)
}
\arguments{
\item{...}{filtering expression to pass to dplyr::filter}

\item{show_hidden}{should hidden tasks be shown (default = FALSE)}
}
\value{
A tibble containing the tasks
}
\description{
View tasks
}
\details{
Displays all tasks, sorted by deadline then priority
}
\examples{
\dontrun{

view_tasks()
}
}
