% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbch.R
\name{work_recover}
\alias{work_recover}
\title{Recover location of missing jobs}
\usage{
work_recover(dirs = getOption("workbch.search"))
}
\arguments{
\item{dirs}{vector of paths to search recursively}
}
\description{
Recover location of missing jobs
}
