% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{work_scan}
\alias{work_scan}
\title{Scan for jobs to create by searching for git repos}
\usage{
work_scan(dir, owner = NULL, status = NULL, priority = NULL)
}
\arguments{
\item{dir}{where to look for git repositories}

\item{owner}{the owner for created jobs (a nick name)}

\item{status}{the status for created jobs}

\item{priority}{the priority of created jobs}
}
\description{
Scan for jobs to create by searching for git repos
}
