% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbch.R
\name{workbch_paths}
\alias{workbch_paths}
\title{Paths known to workbch}
\usage{
workbch_paths(show_hidden = TRUE)
}
\arguments{
\item{show_hidden}{should hidden jobs be included}
}
\value{
A tibble
}
\description{
Paths known to workbch
}
