% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbch.R
\name{workbch_sethome}
\alias{workbch_sethome}
\title{Set location of workbch files}
\usage{
workbch_sethome(path)
}
\arguments{
\item{path}{Path to the folder}
}
\value{
Path to the folder
}
\description{
Set location of workbch files
}
