% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbch.R
\name{workbch_setsearchpath}
\alias{workbch_setsearchpath}
\title{Sets the workbch search paths}
\usage{
workbch_setsearchpath(paths)
}
\arguments{
\item{paths}{Paths to search}
}
\value{
Path to the folder
}
\description{
Sets the workbch search paths
}
