% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_note}
\alias{delete_note}
\title{Delete a note from a job}
\usage{
delete_note(id, jobname = NULL)
}
\arguments{
\item{id}{the id number assigned to the note}

\item{jobname}{the job from which the note should be deleted}
}
\description{
Delete a note from a job
}
\examples{
\dontrun{

delete_note(2, "myjob")
}
}
