% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_task}
\alias{delete_task}
\title{Delete a task from a job}
\usage{
delete_task(id, jobname)
}
\arguments{
\item{id}{the id number assigned to the task}

\item{jobname}{the job from which the task should be deleted}
}
\description{
Delete a task from a job
}
\examples{
\dontrun{

delete_task(2, "myjob")
}
}
