% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_task}
\alias{delete_task}
\title{Delete a task from a job}
\usage{
delete_task(jobname, id)
}
\arguments{
\item{jobname}{the job from which the task should be deleted}

\item{id}{the id number assigned to the task}
}
\description{
Delete a task from a job
}
\examples{
\dontrun{

delete_task("myjob", 2)
}
}
