% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R, R/goto.R
\name{goto_project}
\alias{goto_project}
\title{Navigate to the RStudio project linked with a job}
\usage{
goto_project(name)

goto_project(name)
}
\arguments{
\item{name}{Name of job to open}

\item{name}{Name of job to open}
}
\description{
Navigate to the RStudio project linked with a job

Navigate to the RStudio project linked with a job
}
