% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goto.R
\name{goto_url}
\alias{goto_url}
\title{Navigate to a URL linked to a job}
\usage{
goto_url(name, site)
}
\arguments{
\item{name}{name of the project}

\item{site}{label denoting the site (e.g., "github")}
}
\description{
Navigate to a URL linked to a job
}
