% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goto.R
\name{goto_url}
\alias{goto_url}
\title{Navigate to a URL linked to a job}
\usage{
goto_url(jobname, site)
}
\arguments{
\item{jobname}{name of the job}

\item{site}{label denoting the site (e.g., "github")}
}
\description{
Navigate to a URL linked to a job
}
\examples{
\dontrun{

goto_url("myjob", "github")
}
}
