% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{job_create}
\alias{job_create}
\title{Create a new project}
\usage{
job_create(name, description, owner, status = "active",
  team = character(0), priority = 1, deadline = NA, links = list(),
  notes = list(), tasks = list())
}
\arguments{
\item{name}{name of the project to create}

\item{description}{brief description of the project}

\item{owner}{should be a name or a nickname}

\item{status}{should be "active", "inactive", "complete", "abandoned"}

\item{team}{should be a vector of names/nicknames}

\item{priority}{numeric}

\item{deadline}{a date}

\item{notes}{list of notes}

\item{tasks}{list of tasks}
}
\description{
Create a new project
}
