% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_create.R
\name{job_create}
\alias{job_create}
\title{Create a new job}
\usage{
job_create(jobname = NULL, description = NULL, owner = NULL,
  status = NULL, priority = NULL, tags = NULL, path = NULL)
}
\arguments{
\item{jobname}{name of the job to create}

\item{description}{brief description of the job}

\item{owner}{should be a name or a nickname}

\item{status}{should be "active", "inactive", "complete", "abandoned", "masked"}

\item{priority}{numeric}

\item{tags}{a string containing comma separated list of tags}

\item{path}{path to the job home directory}
}
\value{
Invisibly returns a list containing the parameters for the job
}
\description{
Create a new job
}
\details{
The role of the \code{job_create()} function is to create new workbch job.
It can be called in two ways, interactively or programmatically. To call the
function interactively, R must be in interactive mode and the function should
be called with no arguments specified. When called in this fashion the user
will be presented with a sequence of prompts, asking them to specify each
of the parameters that define a job (e.g., a character string for \code{jobname},
a number for \code{priority}). When used interactively, you do not need to include
quote marks when entering a string: \code{job_create()} will coerce the input to
the appropriate format, and then append the created job to the job file.

When called with programmatically, the user must specify any arguments in the
call to \code{job_create()}. The \code{jobname}, \code{description} and
\code{owner} arguments should be character strings of length 1, and all three
are mandatory. The \code{status} for a job should be one of the following
values: \code{"active"}, \code{"inactive"}, \code{"complete"}, \code{"abandoned"}
or \code{"masked"}. The \code{priority} for
a job should be a positive integer: the intent is that priority 1 is the
highest priority, followed by priority 2, and so one. The \code{tags} for a job
can be specified as a single string, using \code{|} as a separator character
(e.g., \code{tags = "research | statistics"} would create two tags for the job).
Finally, the \code{path} should specify the location of a folder containing the
project files.

Note that, although jobs can also be associated with URLs (e.g., link to a
GitHub repository or a document on Overleaf), the \code{job_create()} function
does not (at this time) allow you to specify URLs. These can be added using
\code{job_modify()}.
}
