% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_project.R
\name{job_edit_team}
\alias{job_edit_team}
\title{Add or remove people from a team}
\usage{
job_edit_team(name, add = NULL, remove = NULL)
}
\arguments{
\item{name}{Name of project}

\item{add}{Vector of names to add to the team}

\item{remove}{Vector of names to remove from the team}
}
\description{
Add or remove people from a team
}
