% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_glimpse.R
\name{job_glimpse}
\alias{job_glimpse}
\title{View the information stored about a job}
\usage{
job_glimpse(jobname = NULL)
}
\arguments{
\item{jobname}{Name of the job to show (defaults to the current job)}
}
\value{
Invisibly returns the job as a list
}
\description{
View the information stored about a job
}
