% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_list.R
\name{job_list}
\alias{job_list}
\title{View a list of all jobs}
\usage{
job_list(search = NULL, select = c("jobname", "owner", "priority",
  "status", "description"))
}
\arguments{
\item{search}{input query used to extract a subset of jobs}

\item{select}{character vector of columns to return}
}
\description{
View a list of all jobs
}
