% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_project.R
\name{job_open}
\alias{job_open}
\title{Open the RStudio project linked to the job}
\usage{
job_open(name)
}
\arguments{
\item{name}{Name of job to open}
}
\description{
Open the RStudio project linked to the job
}
