% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_openurl.R
\name{job_openurl}
\alias{job_openurl}
\title{Navigate to a URL associated with a job}
\usage{
job_openurl(site, jobname = NULL)
}
\arguments{
\item{site}{label denoting the site (e.g., "github")}

\item{jobname}{name of the job}
}
\description{
Navigate to a URL associated with a job
}
