% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_seek.R
\name{job_seek}
\alias{job_seek}
\title{Scans for jobs}
\usage{
job_seek(dirs = getOption("workbch.search"), seek = c("workbch", "git",
  "Rproj"), nesting = FALSE, owner = "", priority = 1,
  status = "active", tags = character(0))
}
\arguments{
\item{dirs}{directories to scan}

\item{seek}{search criterion}

\item{nesting}{allow nested jobs (default = FALSE)}

\item{owner}{default owner for a created job}

\item{priority}{default priority for a created job}

\item{status}{default status for a created job}

\item{tags}{default tags for a created job}
}
\description{
Scans for jobs
}
