% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_seek.R
\name{job_seek}
\alias{job_seek}
\title{Scans for jobs}
\usage{
job_seek(dirs = getOption("workbch.search"), seek = c("workbch", "git",
  "Rproj"), nesting = FALSE, default_owner = "",
  default_priority = 1, default_status = "active",
  default_tags = character(0))
}
\arguments{
\item{dirs}{directories to scan}

\item{seek}{search criterion}

\item{nesting}{allow nested jobs (default = FALSE)}

\item{default_owner}{default owner for a created job}

\item{default_priority}{default priority for a created job}

\item{default_status}{default status for a created job}

\item{default_tags}{default tags for a created job}
}
\description{
Scans for jobs
}
