% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_job}
\alias{make_job}
\title{Create a new job}
\usage{
make_job(jobname, description, owner, status = NULL, team = NULL,
  priority = NULL, deadline = NULL, path = NULL, hidden = NULL)
}
\arguments{
\item{jobname}{name of the job to create}

\item{description}{brief description of the job}

\item{owner}{should be a name or a nickname}

\item{status}{should be "active", "inactive", "complete", "abandoned"}

\item{team}{should be a vector of names/nicknames}

\item{priority}{numeric}

\item{deadline}{a date}

\item{path}{path to the job home directory}

\item{hidden}{hide job (default = FALSE)}
}
\description{
Create a new job
}
