% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goto.R
\name{open_project}
\alias{open_project}
\title{Open the RStudio project linked to the job}
\usage{
open_project(name)
}
\arguments{
\item{name}{Name of job to open}
}
\description{
Open the RStudio project linked to the job
}
