% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_deadline}
\alias{set_deadline}
\title{Sets the deadline for a job}
\usage{
set_deadline(jobname, deadline)
}
\arguments{
\item{jobname}{Name of the job}

\item{deadline}{The new deadline}
}
\description{
Sets the deadline for a job
}
