% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_job}
\alias{set_job}
\title{Set a new job, or modify an existing one}
\usage{
set_job(name, description = NULL, owner = NULL, status = NULL,
  team = NULL, priority = NULL, deadline = NULL, path = NULL,
  hidden = NULL)
}
\arguments{
\item{name}{name of the job to create}

\item{description}{brief description of the job}

\item{owner}{should be a name or a nickname}

\item{status}{should be "active", "inactive", "complete", "abandoned"}

\item{team}{should be a vector of names/nicknames}

\item{priority}{numeric}

\item{deadline}{a date}

\item{path}{path to the project home directory}

\item{hidden}{hide job (default = FALSE)}
}
\description{
Set a new job, or modify an existing one
}
