% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_person}
\alias{set_person}
\title{Sets details for a new person}
\usage{
set_person(fullname, nickname)
}
\arguments{
\item{fullname}{the full name of the person}

\item{nickname}{a short name for the person}
}
\description{
Sets details for a new person
}
