% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_status}
\alias{set_status}
\title{Sets the status of a job}
\usage{
set_status(jobname, status)
}
\arguments{
\item{jobname}{Name of the job}

\item{status}{The new status}
}
\description{
Sets the status of a job
}
