% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_tag}
\alias{set_tag}
\title{Set the tags for a job}
\usage{
set_tag(add = NULL, remove = NULL, jobname = NULL)
}
\arguments{
\item{add}{character vector of tags to add to jobs}

\item{remove}{character vector of tags to remove from jobd}

\item{jobname}{name of job(s) to be edited}
}
\description{
Set the tags for a job
}
\details{
The role of \code{set_tags()} is to...
}
