% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_task_deadline}
\alias{set_task_deadline}
\title{Set the deadline for a taskjob}
\usage{
set_task_deadline(date, id, jobname = NULL)
}
\arguments{
\item{date}{character string to be parsed by lubridate::dmy}

\item{id}{id number of the task to be edited}

\item{jobname}{name of job(s) to be edited}
}
\description{
Set the deadline for a taskjob
}
