% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_url}
\alias{set_url}
\title{Set a URL associated with a job}
\usage{
set_url(name, site, link)
}
\arguments{
\item{name}{name of the job to edit}

\item{site}{string with the site nickname (e.g., "github")}

\item{link}{string with the link to the site}
}
\description{
Set a URL associated with a job
}
\details{
The role of \code{job_edit_url()} is to make it a easier to
change a URL associated with a job.
}
\examples{
\dontrun{

set_url("myjob", "github", "https://github.com/myusername/myrepo")

}
}
