% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewing.R
\name{view_gitstatus}
\alias{view_gitstatus}
\title{Check the git status of projects}
\usage{
view_gitstatus(show_hidden = FALSE)
}
\arguments{
\item{show_hidden}{should hidden jobs be included}
}
\value{
A character vector of names, in alphabetical order
}
\description{
Check the git status of projects
}
