% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_jobnames}
\alias{view_jobnames}
\title{View job names known to workbch}
\usage{
view_jobnames(show_hidden = FALSE)
}
\arguments{
\item{show_hidden}{should hidden jobs be included}
}
\value{
A character vector of names, in alphabetical order
}
\description{
View job names known to workbch
}
