% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_jobs}
\alias{view_jobs}
\title{View a list of jobs}
\usage{
view_jobs(..., show_hidden = FALSE)
}
\arguments{
\item{...}{expression to be passed to dplyr::filter}

\item{show_hidden}{should hidden jobs be included}
}
\description{
View a list of jobs
}
