% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewing.R
\name{view_notes}
\alias{view_notes}
\title{View notes associated with a job}
\usage{
view_notes(name)
}
\arguments{
\item{name}{the job}
}
\value{
a tibble
}
\description{
View notes associated with a job
}
\examples{
\dontrun{

view_notes("myjob")
}
}
