% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view_priorities}
\alias{view_priorities}
\title{View jobs by priority}
\usage{
view_priorities(priority = 1, ..., show_hidden = FALSE)
}
\arguments{
\item{priority}{numeric vector of priorities to display}

\item{...}{expression to be passed to view_jobs}

\item{show_hidden}{should hidden jobs be included}
}
\value{
tibble of jobs
}
\description{
View jobs by priority
}
