% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewing.R
\name{view_projects}
\alias{view_projects}
\title{List the project folders known to workbch}
\usage{
view_projects(show_hidden = FALSE)
}
\arguments{
\item{show_hidden}{should hidden jobs be included}
}
\value{
A tibble
}
\description{
List the project folders known to workbch
}
