% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{workbch_home}
\alias{workbch_home}
\title{Get or set the location of workbch files}
\usage{
workbch_home(path = NULL)
}
\arguments{
\item{path}{Path to the folder}
}
\value{
Path to the folder
}
\description{
Get or set the location of workbch files
}
